% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jarosimilarity.R
\name{jarosimilarity}
\alias{jarosimilarity}
\title{jarosimilarity}
\usage{
jarosimilarity(a, b, th = 0.7, nthreads = 1)
}
\arguments{
\item{a}{a character vector}

\item{b}{a character vector}

\item{th}{threshold for matching. Strings are consired to match if jarosimilarity is > th}

\item{nthreads}{the number of threads to use. If you have multiple cores at your disposal, you can set this to do matching in parallel.
It is \emph{not} set automatically to match the number of cores in your machine.}
}
\value{
Returns a list containing two vectors of integers that indexes to the vectors given as parameters (that is, 'a' and 'b').
Corresponding elements in the vectors describe a match.
}
\description{
Find approximate matches in two (multi-)sets of strings using the JaroSimilarity algorithm.
}
\details{
TBA
}
